pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "arango-orm"
version = "1.0.1"
description = "A SQLAlchemy like ORM implementation for arangodb"
authors = ["Kashif Iftikhar <kashif@compulife.com.pk>"]
license = "GNU General Public License v2 (GPLv2)"
homepage = "https://arango-orm.compulife.com.pk"
repository = "https://github.com/kashifpk/arango-orm"
readme = "README.md"
packages = [{include="arango_orm"}]

[tool.poetry.dependencies]
python = "^3.11"
python-arango = "^7.6.2"
pydantic = "^2.3.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.270"
black = "^23.3.0"
ipython = "^8.15.0"
mkdocs = "^1.5.3"
mkdocs-material = "^9.4.2"
mkdocstrings = {extras = ["python"], version = "^0.23.0"}


[tool.poetry.group.test.dependencies]
pytest = "^7.4.2"

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
src = ["arango_orm"]